add new terraform plugin that running on docker
