+++ src/plugin/terraform/go.mod
@@ -1,0 +1,3 @@
+module dacrane/plugin/terraform
+
+go 1.20
